MILESTONE COMMIT - Angular-Jekyll working well. Code for angular app placed in the _include/app folder and written into html files to enable use of Jekyll (ie Liquid) code on the server.
There are many routes (states) generated on the server.

There is a resolve property on the routes to ensure that data from the ngData folder is made available on the $rootScope prior to the first state loading.

There is a dataservice responsible for providing this data and it is set up to return a promise. It will not hit the server unless it is the first access or it is explicitly told to refresh.

There is a directive that successfully accesses the data placed on the $rootScope.